fix(domain): address domain logic, tracing and import cleanups (H-6, H-7, H-9, H-10, H-14, M-6)
- Evict previousSlug before mutation in ProductServiceImpl (H-6)
- Add round-robin pagination to ProductMediaSweepJob to sweep all pages (H-7)
- Propagate MDC context map to parallel worker threads in OrderCommitCoordinator and PricingServiceImpl (H-9)
- Scope payment idempotency to (idempotencyKey, userId) in PaymentServiceImpl (H-10)
- Guard PaymentOutboxRelay against marking failed batch messages as SENT (H-14)
- Allow null classId in TaxRateServiceImpl.list to return all rates (M-6)
- Replace inline fully qualified types with top-level imports across KeycloakAdminClient, CorsAutoConfigurer, and HttpLoggingFilter

+++ product-service/src/main/java/com/shop/productservice/security/ServiceTokenProvider.java
@@ -47,7 +47,7 @@ public class ServiceTokenProvider {
                                 MediaClientProperties properties) {
         MediaClientProperties.Keycloak kc = properties.keycloak();
         SimpleClientHttpRequestFactory factory = new SimpleClientHttpRequestFactory();
-        int timeoutMs = properties.timeoutMs() > 0 ? properties.timeoutMs() : 5000;
+        int timeoutMs = properties.timeoutMs() > 0 ? (int) properties.timeoutMs() : 5000;
         factory.setConnectTimeout(timeoutMs);
         factory.setReadTimeout(timeoutMs);
         this.restClient = restClientBuilder.requestFactory(factory).build();
